Reformat file structure
